move winit conversions into winit mod
